Fix an integer overflow bug that prevented the creation of page caches with 512 GiBs of memory or more

+++ community/io/src/main/java/org/neo4j/io/pagecache/impl/muninn/PageList.java
@@ -100,7 +100,7 @@ class PageList
         this.memoryManager = memoryManager;
         this.swappers = swappers;
         this.victimPageAddress = victimPageAddress;
-        long bytes = pageCount * META_DATA_BYTES_PER_PAGE;
+        long bytes = ((long) pageCount) * META_DATA_BYTES_PER_PAGE;
         this.baseAddress = memoryManager.allocateAligned( bytes );
         clearMemory( baseAddress, pageCount );
     }
